requirements.txt
# ── Web / HTTP ──────────────────────────────────────────────
requests==2.32.3
urllib3==2.2.2
beautifulsoup4==4.12.3
lxml==5.2.2

# ── Selenium (뉴스 링크 수집) ────────────────────────────────
selenium==4.21.0

# ── 주가 데이터 ──────────────────────────────────────────────
yfinance==0.2.40

# ── 감정 분석 (VADER - 경량 규칙기반) ────────────────────────
vaderSentiment==3.3.2

# ── 데이터 처리 ──────────────────────────────────────────────
pandas==2.2.2

# ── DB ───────────────────────────────────────────────────────
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1

# ── API ──────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1

# ── 설정 / 유틸 ──────────────────────────────────────────────
python-dotenv==1.0.1
pydantic-settings==2.3.1

# ── 테스트 ───────────────────────────────────────────────────
pytest==8.2.2
pytest-cov==5.0.0
httpx==0.27.0          # FastAPI 테스트 클라이언트용